+++ src/aggregate/filehandler_component.py
@@ -1,19 +1,32 @@
 
 
 import os
-
+import glob
 
 #Note: will add try statements when I look at document to figure what exceptions it raises
 
 class FileHandleComponent:
-  def __init__(self, file_path : str) -> None:
+    def __init__(self, file_path : str) -> None:
         self.file_path = file_path
 
-  #My Methods
-  def fix_str_path(self, file_path : str) -> str:
-      return file_path.replace('\\', '/')
 
-  def makedirs(self, file_path : str):
+    #my methods
+    def remove_all_contents_output_frame(self):
+        input_video_path = './frame_extraction/out_frame/'
+
+        if os.path.exists(input_video_path):
+            frames = glob.glob(os.path.join(input_video_path, "*png"))
+
+            for frame in frames:
+                os.remove(frame)
+                self.logger.debug(f"Removed frame: {frame}")
+        else:
+            raise FileNotFoundError("Input file not found.")
+
+    def fix_str_path(self, file_path : str) -> str:
+        return file_path.replace('\\', '/')
+
+    def makedirs(self, file_path : str):
       os.makedirs(file_path, exist_ok=True)
 
       #DOCS: TODO: figure a good way to fix this issue future bug fix
@@ -23,44 +36,44 @@ class FileHandleComponent:
         # if not exist_ok or not path.isdir(name):
         #     raise
 
-  def path_exists(self, file_path : str):
-      return os.path.exists(file_path)
+    def path_exists(self, file_path : str):
+        return os.path.exists(file_path)
 
 
-  #AI genereated for the sake of just having them. Might remove in future if super redundant to reduce code amount
-  def read_file(self) -> str:
-      """Read the entire file content and return as a string."""
-      with open(self.file_path, 'r') as file:
-          return file.read()
+    #AI genereated for the sake of just having them. Might remove in future if super redundant to reduce code amount
+    def read_file(self) -> str:
+        """Read the entire file content and return as a string."""
+        with open(self.file_path, 'r') as file:
+            return file.read()
 
-  def write_file(self, content: str) -> None:
-      """Write the provided content to the file."""
-      with open(self.file_path, 'w') as file:
-          file.write(content)
+    def write_file(self, content: str) -> None:
+        """Write the provided content to the file."""
+        with open(self.file_path, 'w') as file:
+            file.write(content)
 
-  def append_to_file(self, content: str) -> None:
-      """Append the provided content to the file."""
-      with open(self.file_path, 'a') as file:
-          file.write(content)
+    def append_to_file(self, content: str) -> None:
+        """Append the provided content to the file."""
+        with open(self.file_path, 'a') as file:
+            file.write(content)
 
-  def read_lines(self) -> list:
-      """Read the file and return a list of lines."""
-      with open(self.file_path, 'r') as file:
-          return file.readlines()
+    def read_lines(self) -> list:
+        """Read the file and return a list of lines."""
+        with open(self.file_path, 'r') as file:
+            return file.readlines()
 
-  def write_lines(self, lines: list) -> None:
-      """Write a list of lines to the file."""
-      with open(self.file_path, 'w') as file:
-          file.writelines(lines)
+    def write_lines(self, lines: list) -> None:
+        """Write a list of lines to the file."""
+        with open(self.file_path, 'w') as file:
+            file.writelines(lines)
 
-  def file_exists(self) -> bool:
-      """Check if the file exists."""
-      return os.path.isfile(self.file_path)
+    def file_exists(self) -> bool:
+        """Check if the file exists."""
+        return os.path.isfile(self.file_path)
 
-  def delete_file(self) -> None:
-      """Delete the file."""
-      if self.file_exists():
-          os.remove(self.file_path)
+    def delete_file(self) -> None:
+        """Delete the file."""
+        if self.file_exists():
+            os.remove(self.file_path)
